# backend.new/web/swagger/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "\"My App is a continuously updated personal service collection.\"",
        "title": "My App (backend/web/router.go)",
        "contact": {
            "name": "GitHub Discussions",
            "url": "https://github.com/jinyaoMa/my-app/discussions"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/jinyaoMa/my-app/blob/main/LICENSE"
        },
        "version": "1.0.0"
    },
    "basePath": "/api",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Login and get access token",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Password",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/test": {
            "get": {
                "description": "Test pass path",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Test"
                ],
                "summary": "Pass",
                "responses": {
                    "200": {
                        "description": "Pass Path",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerToken": {
            "description": "Authorization Header should contain value started with \"Bearer \" and followed by a JSON Web Token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}